Data Wrangling
Aim: Load the Boston dataset from sklearn library concerns the housing prices in housing city of Boston. The dataset provided has 506 instances with 13 features. Split the data into training and testing sets. Train the model with 80% of the samples and test with the remaining 20%.  
Predict the house prices for testing dataset.